Refactor ContentView: Extract views into separate files
## Summary
- Reduced ContentView from 818 lines to 56 lines (93% reduction)
- Extracted 9 view structs and 3 shared components into organized files
- Improved code maintainability and navigation

## Changes
- **ContentView.swift**: Now contains only the main view switcher
- **Views/IntroView.swift**: App intro and session resume logic
- **Views/PreWorkoutView.swift**: Pre-workout feeling input
- **Views/WorkoutPhaseViews.swift**: Stretching, Warmup, Load, WorkingSet, Rest views
- **Views/CompletionViews.swift**: Review and Done views
- **Views/SharedComponents.swift**: ExerciseHeader, ActionBar, DecisionBadge

## Testing
- ✅ Built successfully on iOS Simulator (iPhone 17)
- ✅ Verified UI rendering and navigation flow
- ✅ All views display correctly with proper state management

## Related Issues
Addresses codebase cleanup initiative - maintaining lean implementations as the codebase grows.
